drivers: serial: ra: fix issues with interrupt driven operation
In order to initiate an interrupt driven transmisson from the UART on an
RA Microcontroller the TE and TIE bits have to be set simultaneously. If
they are not then the initial transmit interrupt, required to begin the
transmit process, is never generated. This patch resolve the issue and
also fixes another minor bug with transmission complete detection.

Signed-off-by: Ian Morris <ian.d.morris@outlook.com>